reorganize components and customize contact.tsx
