+++ components.json
@@ -11,7 +11,7 @@
     "prefix": ""
   },
   "aliases": {
-    "components": "@/components/ui/",
+    "components": "@/components/shadcn/",
     "utils": "@/lib/"
   }
 }
