+++ src/components/mode-toggle.tsx
@@ -1,12 +1,12 @@
 import { Moon, Sun } from 'lucide-react'
 
-import { Button } from '@/components/ui/button'
+import { Button } from '@/components/shadcn/ui/button'
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from '@/components/ui/dropdown-menu'
+} from '@/components/shadcn/ui/dropdown-menu'
 import { useTheme } from '@/components/theme-provider'
 
 export function ModeToggle() {
